+++ lib/style.ex
@@ -1,4 +1,4 @@
-# Copyright 2023 Adobe. All rights reserved.
+# Copyright 2024 Adobe. All rights reserved.
 # This file is licensed to you under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License. You may obtain a copy
 # of the License at http://www.apache.org/licenses/LICENSE-2.0
@@ -63,6 +63,9 @@ defmodule Styler.Style do
     |> Zipper.traverse(fn zipper -> Zipper.update(zipper, &Macro.update_meta(&1, meta_fun)) end)
     |> Zipper.root()
   end
+
+  # useful for comparing AST without meta (line numbers, etc) interfering
+  def without_meta(ast), do: update_all_meta(ast, fn _ -> nil end)
 
   @doc """
   Returns the current node (wrapped in a `__block__` if necessary) if it's a valid place to insert additional nodes
@@ -152,5 +155,83 @@ defmodule Styler.Style do
       end
     end)
     |> Enum.sort_by(& &1.line)
+  end
+
+  @doc """
+  Takes a list of nodes and clumps them up, setting `end_of_expression: [newlines: x]` to 1 for all but the final node,
+  which gets 2 instead, (hopefully!) creating an empty line before whatever follows.
+  """
+  def reset_newlines([]), do: []
+  def reset_newlines(nodes), do: reset_newlines(nodes, [])
+
+  def reset_newlines([node], acc), do: Enum.reverse([set_newlines(node, 2) | acc])
+  def reset_newlines([node | nodes], acc), do: reset_newlines(nodes, [set_newlines(node, 1) | acc])
+
+  defp set_newlines({directive, meta, children}, newline) do
+    updated_meta = Keyword.update(meta, :end_of_expression, [newlines: newline], &Keyword.put(&1, :newlines, newline))
+    {directive, updated_meta, children}
+  end
+
+  @doc """
+  "Fixes" the line numbers of nodes who have had their orders changed via sorting or other methods.
+  This "fix" siply ensures that comments don't get wrecked as part of us moving AST nodes willy-nilly.
+
+  The fix is rather naive, and simply enforces the following property on the code:
+  A given node must have a line number less than the following node.
+  Et voila! Comments behave much better.
+
+  ## In Detail
+
+  For example, given document
+
+    1: defmodule ...
+    2: alias B
+    3: # this is foo
+    4: def foo ...
+    5: alias A
+
+  Sorting aliases the ast node for  would put `alias A` (line 5) before `alias B` (line 2).
+
+    1: defmodule ...
+    5: alias A
+    2: alias B
+    3: # this is foo
+    4: def foo ...
+
+  Elixir's document algebra would then encounter `line: 5` and immediately dump all comments with `line <= 5`,
+  meaning after running through the formatter we'd end up with
+
+    1: defmodule
+    2: # hi
+    3: # this is foo
+    4: alias A
+    5: alias B
+    6:
+    7: def foo ...
+
+  This function fixes that by seeing that `alias A` has a higher line number than its following sibling `alias B` and so
+  updates `alias A`'s line to be preceding `alias B`'s line.
+
+  Running the results of this function through the formatter now no longer dumps the comments prematurely
+
+    1: defmodule ...
+    2: alias A
+    3: alias B
+    4: # this is foo
+    5: def foo ...
+  """
+  def fix_line_numbers(nodes, nil), do: fix_line_numbers(nodes, 999_999)
+  def fix_line_numbers(nodes, {_, meta, _}), do: fix_line_numbers(nodes, meta[:line])
+  def fix_line_numbers(nodes, max), do: nodes |> Enum.reverse() |> do_fix_lines(max, [])
+
+  defp do_fix_lines([], _, acc), do: acc
+
+  defp do_fix_lines([{_, meta, _} = node | nodes], max, acc) do
+    line = meta[:line]
+
+    # the -2 is just an ugly hack to leave room for one-liner comments and not hijack them.
+    if line > max,
+      do: do_fix_lines(nodes, max, [shift_line(node, max - line - 2) | acc]),
+      else: do_fix_lines(nodes, line, [node | acc])
   end
 end